Improve ignoredErrors
This implementation uses a slice of errors compared to a hash. This
could impact performance if the ignoredErrors list is large, but I
suspect it's typically not.

Switching to a slice means the error only needs to be comparable.
Comparatively using a hash meant the error had to be hashable, and
some types that implement error are not. These cases were likely
silently ignored before go1.12.

Fixes #26