Fix broken recover in nested call and wait for panic to reach Rollbar
My tests showed that Hook.ReportPanic was not able to recover any
panic that was trying to be handled via defer ReportPanic().
Inlining the recover call into the top-level function fixed that.
After that it made no sense to me to keep Hook.ReportPanic.

Fredrik just helped rebase Marc's changes.

Co-authored-by: Fredrik Blomqvist <fredrik.blomqvist.95@gmail.com>