Deliver non-terminal errors to rollbar asynchronously
The message delivery code in rollrus is already written to wait for
reporting panics and fatal errors. However, it also used the rollbar
SyncTransport, which always synchronously reports and `Wait` is a noop.

Unfortunately, rollbar's AsyncTransport does not have clear semantics
and has unresolved issues which can lead to panics[1].

This introduces a new buffered transport which wraps the SyncTransport
and installs it by default for all hooks. The value of
`rollbar.DefaultBuffer` is used to determine the buffer size.

In the buffered transport, it is safe to arbitrarily interleave calls to
`Send` and `Wait`. `Wait` will never return before a prior call to
`Send` unless the transport is closed.

  [1]: https://github.com/rollbar/rollbar-go/issues/68